agent/requirements.txt
# AI Trading Agent Dependencies

# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai              # For OpenRouter (OpenAI-compatible)
langchain-mcp-adapters        # MCP client for LangGraph

# Database
sqlmodel>=0.0.16
aiosqlite                     # Async SQLite

# Messaging
python-telegram-bot>=21.0

# UI
streamlit>=1.30.0
rich>=13.0.0
typer>=0.9.0

# Utils
python-dotenv>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
aiohttp>=3.9.0
dspy-ai
numpy>=1.24.0
TA-Lib>=0.4.0
